refactor: util 디렉터리 분리
cardBrand는 domain으로, 나머지는 utils 디렉터리로 변경
