+++ hooks/src/lib/domain/cardBrand/cardBrand.constant.ts
